requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.16.1
identify==2.6.1
idna==3.10
nodeenv==1.9.1
oauthlib==3.2.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.7.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
